pyproject.toml
[tool.poetry]
name = "telod"
version = "0.1.0"
description = ""
authors = ["Voronwe Sul <mail@voronwe.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
psycopg2-binary = "^2.9.10"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
lxml = "^5.3.0"
python-telegram-bot = "^21.6"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
pre-commit = "^4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
